pyproject.toml
[project]
name = "hacklondon2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.10",
    "imblearn>=0.0",
    "itsdangerous>=2.2.0",
    "llvmlite>=0.44.0",
    "numba>=0.61.0",
    "numpy>=2.1.3",
    "openai>=1.65.2",
    "pandas>=2.2.3",
    "polars>=1.23.0",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "setuptools>=75.8.2",
    "shap>=0.46.0",
    "sqlmodel>=0.0.23",
    "starlette>=0.46.0",
    "tsfresh>=0.21.0",
    "uvicorn>=0.34.0",
]
